ipyauth/requirements.txt
requests
python-dotenv>=0.8.2
ipywidgets>=7.1.2
